requirements.txt
aiohttp
astroid
async-timeout
attrs
cffi
chardet
colorama
cryptography
discord.py
flask
ffmpeg
idna
isort
lazy-object-proxy
mccabe
multidict
pycparser
pylint
pymongo
PyNaCl
pyspnego
six
smbprotocol
toml
typing_extensions
wrapt
yarl
yt-dlp
